:sparkles: 리스트 조회 StatusType을 수정을 적용한 예외처리 구현
- requestParam name이 status가 아닌 경우
- requestParam Type이 complete, incomplete가 아닌 경우
- StringToStatusType 컨버터 구현
