+++ noriter/src/main/java/com/codewarts/noriter/article/domain/type/StatusType.java
@@ -1,12 +1,7 @@
 package com.codewarts.noriter.article.domain.type;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
 public enum StatusType {
 
-    SOLVED, UNSOLVED, COMPLETE, INCOMPLETE;
+    COMPLETE, INCOMPLETE;
 
 }
