+++ noriter/src/main/java/com/codewarts/noriter/article/repository/StudyRepository.java
@@ -1,6 +1,7 @@
 package com.codewarts.noriter.article.repository;
 
 import com.codewarts.noriter.article.domain.Study;
+import com.codewarts.noriter.article.domain.type.StatusType;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,7 +11,7 @@ import org.springframework.data.repository.query.Param;
 public interface StudyRepository extends JpaRepository<Study, Long> {
 
     @Query("select s from Study s where s.status = :status")
-    List<Study> findStudyByCompleted(@Param("status") String status);
+    List<Study> findStudyByCompleted(@Param("status") StatusType status);
 
     @Query("select s from Study s where s.id =:id and s.writer.id =:writerId")
     Optional<Study> findByStudyIdAndWriterId(@Param("id") Long id, @Param("writerId") Long writerId);
