+++ noriter/src/main/java/com/codewarts/noriter/common/config/WebConfig.java
@@ -2,13 +2,16 @@ package com.codewarts.noriter.common.config;
 
 
 import com.codewarts.noriter.auth.jwt.JwtProvider;
+import com.codewarts.noriter.common.converter.StatusTypeConverter;
 import com.codewarts.noriter.interceptor.AuthVerificationInterceptor;
 import io.netty.resolver.DefaultAddressResolverGroup;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
 import org.springframework.http.client.reactive.ReactorClientHttpConnector;
 import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import reactor.netty.http.client.HttpClient;
@@ -37,5 +40,19 @@ public class WebConfig implements WebMvcConfigurer {
         registry.addInterceptor(new AuthVerificationInterceptor(jwtProvider)).order(1)
             .addPathPatterns("/community/playground", "/community/playground/{id}",
                 "/community/gathering", "/community/gathering/{id}");
+    }
+
+    @Override
+    public void addFormatters(FormatterRegistry registry) {
+        registry.addConverter(new StatusTypeConverter.StringToStatusTypeConverter());
+    }
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+            .allowedOrigins("*")
+            .exposedHeaders("*")
+            .allowedHeaders("*")
+            .allowedMethods("GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS");
     }
 }
