+++ noriter/src/main/java/com/codewarts/noriter/exception/type/CommonExceptionType.java
@@ -7,7 +7,9 @@ import org.springframework.http.HttpStatus;
 public enum CommonExceptionType implements ExceptionType {
 
 
-    INVALID_REQUEST("INVALID001", "잘못된 요청입니다.", HttpStatus.BAD_REQUEST);
+    INVALID_REQUEST("INVALID001", "잘못된 요청입니다.", HttpStatus.BAD_REQUEST),
+    INCORRECT_REQUEST_PARAM("INVALID002", "올바르지 않은 requestParam입니다.", HttpStatus.BAD_REQUEST),
+    INCORRECT_REQUEST_PARAM_TYPE("INVALID003", "올바르지 않은 requestParam 타입입니다.", HttpStatus.BAD_REQUEST);
     private final String errorCode;
     private final String message;
     private final HttpStatus httpStatus;
